Parse Tree 생성: 객체->포인터로 바꿈. Tuple 구조체 변경. 동적할당으로 다 변경

+++ scanner.cpp
@@ -49,7 +49,7 @@ bool isReservedWord(string token) { //is reserved word?
 
 bool isMultiToken(string token, vector<string>& real_tokens) {
 	cout << "isMultiToken?" << endl;
-	int size = terminals.size();
+	int size = (int)terminals.size();
 	for (int i = 0; i < size; i++) {
 		const string& s = terminals[i];
 		cout << s << " check " << endl;
@@ -62,7 +62,7 @@ bool isMultiToken(string token, vector<string>& real_tokens) {
 }
 
 void addRealToken(string token, vector<string>& real_tokens) {
-	int size = token.size();
+	int size = (int)token.size();
 	int idx = 0;
 	while (idx < size) {
 		string tok;
